switched installation from setup.py to setup.cfg + fixed tests
