--- MANIFEST.in
@@ -1,8 +1,0 @@
-include *.py
-include pytest.ini
-include LICENSE
-include requirements.txt
-include README.md
-recursive-include test *.py
-recursive-include utils *.py
-include test/data/testfile
