+++ codescanner_analysis/__init__.py
@@ -1,0 +1,25 @@
+from codescanner_analysis import utils
+from codescanner_analysis.byte_plot import BytePlot
+from codescanner_analysis.codescan_interface import CodescanInterface
+from codescanner_analysis.codescanner_analysis import CodescannerAnalysisData
+from codescanner_analysis.color_map import ColorMap
+from codescanner_analysis.comparison_analysis import ComparisonAnalysis
+from codescanner_analysis import file_header_parser
+from codescanner_analysis.overlay_plot import OverlayPlot
+from codescanner_analysis.plot_base import PlotBase
+
+
+# major, minor, micro, release, serial
+__VERSION_INFO__ = (1, 2, 3, 'final', 0)
+__VERSION__ = '%d.%d' % (__VERSION_INFO__[0], __VERSION_INFO__[1])
+__all__ = [
+    __VERSION__,
+    BytePlot,
+    CodescannerAnalysisData,
+    CodescanInterface,
+    ColorMap,
+    ComparisonAnalysis,
+    file_header_parser,
+    OverlayPlot,
+    PlotBase,
+]
