+++ codescanner_analysis/res/lib/languages/ppc_be.lang
