+++ codescanner_analysis/utils/file_to_array.py
