+++ setup.cfg
@@ -1,0 +1,23 @@
+[metadata]
+name = codescanner_analysis
+version = attr: codescanner_analysis.__VERSION__
+description = Analysis module to classify regions of arbitrary binaries
+author = Fraunhofer FKIE CA&D, Viviane Zwanger, Henning Braun
+author_email = viviane.zwanger@fkie.fraunhofer.de, henning.braun@fkie.fraunhofer.de
+url = https://github.com/fkie-cad/Codescanner
+keywords = codescanner, code regions
+license = Fraunhofer non-commercial License "Codescanner"
+
+[options]
+packages =
+    codescanner_analysis
+    codescanner_analysis.utils
+install_requires =
+    matplotlib
+    numpy
+
+[options.package_data]
+codescanner_analysis =
+    res/lib/codescan
+    res/lib/*.so
+    res/lib/languages/*.lang
